chore: rename service name
